Adding support for user role.
